apis/flask-dockerops-api/requirements.txt
Flask
Flask-RESTful
tornado
gevent
setproctitle
requests
python-etcd
SpliceURL>=0.3
paramiko
